refactor(tui): move input box render into agent_tui_kit (Ctx slice)
