+++ crates/tui/src/render/input.rs
@@ -1,85 +1,18 @@
-use ratatui::{
-    Frame,
-    layout::{Alignment, Rect},
-    style::{Color, Style},
-    text::{Line, Span, Text},
-    widgets::{Block, Borders, Clear, Paragraph},
-};
-use unicode_width::{UnicodeWidthChar, UnicodeWidthStr};
+//! Input box / command line — app-layer wrappers.
+//!
+//! The rendering moved verbatim into `agent_tui_kit::render::input` (pure
+//! `&RenderCtx`). This module keeps the `&mut App` prepare phase (caret-scroll
+//! adjustment, `[Cancel]` hit area, voice button hit area) plus the
+//! App-integration tests that render through the wrapper.
 
-use super::slash_style::{skill_name_set, style_input_skill_line};
+use ratatui::{Frame, layout::Rect};
+
+use agent_tui_kit::render::input as kit_input;
+
+use super::slash_style::skill_name_set;
 use crate::widgets::state::{App, InputMode, VoicePhase};
 
-/// Soft-wrap a logical line into display-line slices no wider than
-/// `max_width` columns, splitting at character boundaries.
-///
-/// The renderer draws exactly these rows (Paragraph stays unwrapped), so the
-/// caret computed from the same function lands where the text is visible.
-pub(crate) fn wrap_line(line: &str, max_width: usize) -> Vec<&str> {
-    let mut rows = Vec::new();
-    if max_width == 0 {
-        rows.push(line);
-        return rows;
-    }
-    let mut start = 0;
-    let mut width = 0;
-    for (byte_idx, ch) in line.char_indices() {
-        let w = UnicodeWidthChar::width(ch).unwrap_or(0);
-        if width + w > max_width && width > 0 {
-            rows.push(&line[start..byte_idx]);
-            start = byte_idx;
-            width = 0;
-        }
-        width += w;
-    }
-    rows.push(&line[start..]);
-    rows
-}
-
-/// Display (row, column) of a caret sitting at logical `cursor_col` within
-/// `line`, after soft-wrapping at `max_width`. The caret is placed at the end
-/// of the prefix's last display row.
-fn caret_in_wrapped(line: &str, max_width: usize, cursor_col: usize) -> (usize, usize) {
-    let mut acc = 0;
-    let mut prefix_len = 0;
-    for (byte_idx, ch) in line.char_indices() {
-        if acc >= cursor_col {
-            break;
-        }
-        let w = UnicodeWidthChar::width(ch).unwrap_or(0);
-        if acc + w > cursor_col {
-            break;
-        }
-        acc += w;
-        prefix_len = byte_idx + ch.len_utf8();
-    }
-    let wrapped = wrap_line(&line[..prefix_len], max_width);
-    let row = wrapped.len().saturating_sub(1);
-    let col = UnicodeWidthStr::width(wrapped.last().copied().unwrap_or(""));
-    (row, col)
-}
-
-/// Render command-line input (Palette mode).
-pub(crate) fn render_command_line(frame: &mut Frame, area: Rect, app: &App) {
-    let content = app.cmd_line.clone();
-    let input_para = Paragraph::new(content)
-        .style(
-            Style::default()
-                .fg(app.theme.input_box_fg)
-                .bg(app.theme.input_box_bg),
-        )
-        .block(
-            Block::default()
-                .borders(Borders::ALL)
-                .title(app.msgs().command_title),
-        );
-    frame.render_widget(input_para, area);
-    let cmd_width = UnicodeWidthStr::width(app.cmd_line.as_str()) as u16;
-    let cursor_pos = (area.x + 2 + cmd_width).min(area.x + area.width - 2);
-    frame.set_cursor_position((cursor_pos, area.y + 1));
-}
-
-/// Render the main input box (Insert mode), or delegate to command-line rendering.
+/// Render the input box (insert mode) or command line (palette mode).
 pub(crate) fn render_input_box(frame: &mut Frame, area: Rect, app: &mut App) {
     if app.input_mode == InputMode::Palette {
         render_command_line(frame, area, app);
@@ -87,18 +20,15 @@ pub(crate) fn render_input_box(frame: &mut Frame, area: Rect, app: &mut App) {
     }
 
     // Codex-style queued messages: the hint + "↳ message" rows sit above the
-    // bordered input box, which shrinks to the remaining height. The pending
-    // block paints its own background (TUI invariant: no shadow residue).
-    let pending_lines = app.pending_display_lines();
-    let area = if pending_lines == 0 {
-        app.set_cancel_button_area(Rect::default());
+    // bordered input box, which shrinks to the remaining height.
+    let pending_lines = if app.pending_messages.is_empty() {
+        0
+    } else {
+        (1 + app.pending_messages.len()).min(4) as u16
+    };
+    let input_area = if pending_lines == 0 {
         area
     } else {
-        render_pending_block(
-            frame,
-            Rect::new(area.x, area.y, area.width, pending_lines),
-            app,
-        );
         Rect::new(
             area.x,
             area.y + pending_lines,
@@ -107,237 +37,30 @@ pub(crate) fn render_input_box(frame: &mut Frame, area: Rect, app: &mut App) {
         )
     };
 
-    // Logical rows are separated by explicit `\n`; the caret line/column is
-    // computed on the raw input, then mapped through soft-wrapping below.
-    let mut cursor_line = 0;
-    let mut cursor_col = 0;
-    for (i, c) in app.input.char_indices() {
-        if i >= app.input_cursor {
-            break;
-        }
-        if c == '\n' {
-            cursor_line += 1;
-            cursor_col = 0;
-        } else {
-            cursor_col += UnicodeWidthChar::width(c).unwrap_or(0);
-        }
-    }
-
-    let inner_width = area.width.saturating_sub(2).max(1) as usize;
-    let lines: Vec<&str> = app.input.split('\n').collect();
-
-    // Caret position after soft-wrapping: rows before the caret's logical
-    // line plus the caret's row within that line.
-    let cursor_logical = lines.get(cursor_line).copied().unwrap_or("");
-    let (caret_row_in_line, caret_col_in_line) =
-        caret_in_wrapped(cursor_logical, inner_width, cursor_col);
-    let prior_display_rows: usize = lines[..cursor_line]
-        .iter()
-        .map(|line| wrap_line(line, inner_width).len())
-        .sum();
-    let cursor_display_line = prior_display_rows + caret_row_in_line;
-
-    // Every display row the input occupies after soft-wrapping.
-    let display_lines: Vec<&str> = lines
-        .iter()
-        .flat_map(|line| wrap_line(line, inner_width))
-        .collect();
-
-    let visible_lines = area.height.saturating_sub(2) as usize;
-
+    // Prepare: keep the caret visible (mutable — the kit render pass is pure).
+    let inner_width = input_area.width.saturating_sub(2).max(1) as usize;
+    let (cursor_display_line, _) =
+        kit_input::caret_display_line(&app.input, app.input_cursor, inner_width);
+    let visible_lines = input_area.height.saturating_sub(2) as usize;
     if cursor_display_line < app.input_scroll as usize {
         app.input_scroll = cursor_display_line as u16;
     } else if cursor_display_line >= app.input_scroll as usize + visible_lines {
         app.input_scroll = (cursor_display_line - visible_lines + 1) as u16;
     }
 
-    let line_stats = if app.input.is_empty() {
-        None
-    } else {
-        Some((display_lines.len(), app.input.chars().count()))
-    };
+    // Prepare: voice button hit area (app-layer extension).
+    update_voice_button_area(app, input_area);
 
-    let start = app.input_scroll as usize;
-    let end = (start + visible_lines).min(display_lines.len());
-    let placeholder_mode = app.input.is_empty();
-
-    let display: Text<'static> = if placeholder_mode {
-        Text::from(Span::styled(
-            app.msgs().input_box_placeholder.to_string(),
-            Style::default()
-                .fg(Color::Rgb(100, 100, 120))
-                .bg(app.theme.input_box_bg),
-        ))
-    } else {
-        let skill_names = skill_name_set(&app.skills_data);
-        let styled_lines: Vec<Line<'static>> = display_lines[start..end]
-            .iter()
-            .map(|line| {
-                style_input_skill_line(line, &skill_names, &app.theme).unwrap_or_else(|| {
-                    Line::from(Span::styled(
-                        (*line).to_string(),
-                        Style::default()
-                            .fg(app.theme.input_box_fg)
-                            .bg(app.theme.input_box_bg),
-                    ))
-                })
-            })
-            .collect();
-        Text::from(styled_lines)
-    };
-
-    let bottom_title = line_stats
-        .map(|(total_lines, total_chars)| format!(" 📝 {total_lines}L · {total_chars}chars "))
-        .unwrap_or_default();
-
-    // Determine border color: accent when focused (insert mode), normal otherwise
-    let border_color = if app.input_mode == InputMode::Insert {
-        app.theme.accent
-    } else {
-        app.theme.border
-    };
-
-    // Left title + centered voice as separate Block titles so the top border
-    // stays visible between them (space-padding with a bg would eat the line).
-    let mut block = Block::default()
-        .borders(Borders::ALL)
-        .border_type(app.theme.block_border_type())
-        .border_style(Style::default().fg(border_color))
-        .title(Span::raw(app.msgs().input_box_title))
-        .title_bottom(bottom_title);
-    if let Some((voice_label, voice_style)) = voice_title(app) {
-        block = block
-            .title(Line::from(Span::styled(voice_label, voice_style)).alignment(Alignment::Center));
-    }
-
-    let input_para = Paragraph::new(display)
-        .style(Style::default().bg(app.theme.input_box_bg))
-        .block(block);
-    frame.render_widget(Clear, area);
-    frame.render_widget(input_para, area);
-    update_voice_button_area(app, area);
-
-    let cursor_x =
-        (area.x + 1 + caret_col_in_line as u16).min(area.x + area.width.saturating_sub(2));
-    let cursor_y = area.y + 1 + (cursor_display_line - app.input_scroll as usize) as u16;
-    frame.set_cursor_position((cursor_x, cursor_y));
-}
-
-/// Truncate a string to at most `max` display columns, appending `…` when cut.
-fn truncate_to_width(s: &str, max: usize) -> String {
-    if UnicodeWidthStr::width(s) <= max {
-        return s.to_string();
-    }
-    let mut out = String::new();
-    let mut width = 0;
-    for c in s.chars() {
-        let cw = UnicodeWidthChar::width(c).unwrap_or(0);
-        if width + cw > max.saturating_sub(1) {
-            break;
-        }
-        out.push(c);
-        width += cw;
-    }
-    out.push('…');
-    out
-}
-
-/// Codex-style pending block: a hint line plus one `↳ message` row per queued
-/// message, drawn above the input box while the agent is busy. The hint row
-/// carries a clickable `[Cancel]` button (mouse) — the only way to drop the
-/// queued messages; they otherwise auto-submit when the current task finishes.
-fn render_pending_block(frame: &mut Frame, area: Rect, app: &mut App) {
-    if area.height == 0 || area.width == 0 {
-        return;
-    }
-    let inner_width = area.width.saturating_sub(2).max(1) as usize;
-    let cancel_label = app.msgs().pending_cancel_btn;
-    let cancel_width = UnicodeWidthStr::width(cancel_label) + 2; // "[label]"
-    let can_show_cancel = !app.pending_messages.is_empty() && inner_width >= cancel_width + 30;
-    let hint = app.msgs().pending_submit_hint;
-    let hint_max = if can_show_cancel {
-        inner_width.saturating_sub(cancel_width).saturating_sub(1)
-    } else {
-        inner_width
-    };
-    let hint_text = truncate_to_width(hint, hint_max);
-    let hint_width = UnicodeWidthStr::width(hint_text.as_str());
-    let cancel_area = if can_show_cancel {
-        Rect::new(
-            area.x + hint_width as u16 + 1,
-            area.y,
-            cancel_width as u16,
-            1,
-        )
-    } else {
-        Rect::default()
-    };
+    let ctx = app.render_ctx();
+    let skill_names = skill_name_set(&app.skills_data);
+    let cancel_area = kit_input::render_input_box(frame, area, &ctx, &skill_names);
     app.set_cancel_button_area(cancel_area);
-
-    let mut lines = vec![Line::from(Span::styled(
-        hint_text,
-        Style::default()
-            .fg(app.theme.warning)
-            .bg(app.theme.input_box_bg),
-    ))];
-    if !cancel_area.is_empty() {
-        lines[0].spans.push(Span::styled(
-            " ",
-            Style::default().bg(app.theme.input_box_bg),
-        ));
-        lines[0].spans.push(Span::styled(
-            format!("[{cancel_label}]"),
-            Style::default()
-                .fg(app.theme.warning)
-                .bg(app.theme.input_box_bg),
-        ));
-    }
-    for pending in app.pending_messages.iter().take(3) {
-        let text = truncate_to_width(&format!("↳ {}", pending.display), inner_width);
-        lines.push(Line::from(Span::styled(
-            text,
-            Style::default()
-                .fg(app.theme.input_box_fg)
-                .bg(app.theme.input_box_bg),
-        )));
-    }
-    frame.render_widget(
-        Paragraph::new(Text::from(lines)).style(Style::default().bg(app.theme.input_box_bg)),
-        area,
-    );
 }
 
-fn voice_title(app: &App) -> Option<(String, Style)> {
-    if matches!(app.voice.phase, VoicePhase::Disabled) {
-        return None;
-    }
-    let label = voice_button_label(app);
-    if label.is_empty() {
-        return None;
-    }
-    let style = match app.voice.phase {
-        VoicePhase::Idle => Style::default().fg(app.theme.accent),
-        VoicePhase::Recording { .. } => Style::default().fg(app.theme.warning),
-        VoicePhase::Transcribing => Style::default().fg(Color::Rgb(120, 120, 140)),
-        VoicePhase::Disabled => Style::default(),
-    };
-    Some((label, style))
-}
-
-fn voice_button_label(app: &App) -> String {
-    match app.voice.phase {
-        VoicePhase::Disabled => String::new(),
-        VoicePhase::Idle => app.msgs().voice_idle.to_string(),
-        VoicePhase::Recording { started_at } => {
-            let elapsed = started_at.elapsed();
-            format!(
-                "⏹ {:02}:{:02}",
-                elapsed.as_secs() / 60,
-                elapsed.as_secs() % 60
-            )
-        }
-        VoicePhase::Transcribing => app.msgs().voice_transcribing.to_string(),
-    }
+/// Render command-line input (Palette mode).
+pub(crate) fn render_command_line(frame: &mut Frame, area: Rect, app: &App) {
+    let ctx = app.render_ctx();
+    kit_input::render_command_line(frame, area, &ctx);
 }
 
 fn update_voice_button_area(app: &mut App, area: Rect) {
@@ -345,8 +68,8 @@ fn update_voice_button_area(app: &mut App, area: Rect) {
         app.voice.set_button_area(Rect::default());
         return;
     }
-    let label = voice_button_label(app);
-    let width = UnicodeWidthStr::width(label.as_str()) as u16;
+    let label = app.voice_button_label();
+    let width = unicode_width::UnicodeWidthStr::width(label.as_str()) as u16;
     if width == 0 {
         app.voice.set_button_area(Rect::default());
         return;
@@ -366,9 +89,10 @@ mod render_tests {
 
     use super::{
         super::test_harness::{buffer_text, make_app},
-        caret_in_wrapped, render_input_box, wrap_line,
+        render_input_box,
     };
     use crate::widgets::state::{SkillEntry, VoicePhase, VoiceState};
+    use agent_tui_kit::render::input::{caret_in_wrapped, truncate_to_width, wrap_line};
 
     #[test]
     fn input_box_renders_multiline_content() {
@@ -527,12 +251,12 @@ mod render_tests {
 
     #[test]
     fn truncate_to_width_appends_ellipsis_for_cjk() {
-        assert_eq!(super::truncate_to_width("abc", 5), "abc");
-        assert_eq!(super::truncate_to_width("abcdef", 4), "abc…");
+        assert_eq!(truncate_to_width("abc", 5), "abc");
+        assert_eq!(truncate_to_width("abcdef", 4), "abc…");
         // CJK wide chars count as two columns; truncation targets `max` cols.
-        assert_eq!(super::truncate_to_width("中文abc", 5), "中文…");
-        assert_eq!(super::truncate_to_width("中文abc", 6), "中文a…");
-        assert_eq!(super::truncate_to_width("中文abc", 7), "中文abc");
+        assert_eq!(truncate_to_width("中文abc", 5), "中文…");
+        assert_eq!(truncate_to_width("中文abc", 6), "中文a…");
+        assert_eq!(truncate_to_width("中文abc", 7), "中文abc");
     }
 
     #[test]
