+++ crates/tui/src/render/mod.rs
@@ -19,8 +19,9 @@ pub(crate) mod util {
     pub(crate) use agent_tui_kit::render::util::*;
 }
 
+pub(super) use agent_tui_kit::render::input::wrap_line;
 pub(super) use bar::{render_bottom_bar, render_status_bar};
-pub(super) use input::{render_input_box, wrap_line};
+pub(super) use input::render_input_box;
 pub(super) use layout::render_main_area;
 pub(super) use popups::{
     command_palette::render_command_palette, file_picker::render_file_picker,
